pyproject.toml
[tool.poetry]
name = "stratton-oakmont"
version = "0.1.0"
description = ""
authors = ["Vasanth-12 <vasanth.dpalanisamy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-sam-cli = "^1.105.0"
numpy = "^1.26.2"
opencv-python = "^4.9.0.80"
pillow = "^10.1.0"
uvicorn = "^0.25.0"
pymongo = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
